fix(historyTask): arreglado move de tarea

+++ src/main/java/com/spring/Service/HistoryTaskService.java
@@ -68,7 +68,7 @@ public class HistoryTaskService extends AbstractService {
 		this.validateBoxPrivileges(destiny.getWorkspace().getSprint(), destiny.getWorkspace());
 		Task task = serviceTask.findOne(dto.getTask());
 		Column origin = task.getColumn();
-		if(origin.getName() == "Done") {
+		if(origin != null && origin.getName().equals("Done")) {
 			HistoryTask historyTask = this.repository.findByTaskAndDestiny(task, origin);
 			this.repository.delete(historyTask);
 		}
